refactor(kernel): phase 3b — extract cron tick loop closure to kernel::cron_tick
The cron scheduler tick loop was historically the longest closure in
mod.rs (~528 lines, the landing zone for #4683 et al.). Lift the body
out of `spawn_logged("cron_scheduler", async move { … })` into a free
`pub(super) async fn run_cron_scheduler_loop(kernel: Arc<LibreFangKernel>)`
in `crates/librefang-kernel/src/kernel/cron_tick.rs`.

Behaviour-preserving — body moved byte-for-byte; only the outer wrapper
changed (closure → free fn). Captured state was just `kernel`, which
becomes the single fn parameter; per-tick `cron_sem` and per-job clones
are still constructed inside the loop body, unchanged.

mod.rs drops by ~528 lines (now ~16.2k). Three Phase 2 re-exports
(`cron_deliver_response`, `cron_fan_out_targets`, `cron_script_wake_gate`)
that mod.rs introduced for the inline closure are no longer needed —
the only call sites moved with the loop — so they're gone from mod.rs's
`use` block.

Verified: cargo check / clippy -- -D warnings / cargo check --tests
all green.
